requirements.txt
langchain>=0.1.12
#langchain-huggingface>=0.2.0
langchain-openai>=0.0.8
#huggingface-hub[inference]
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.27.1
#streamlit==1.32.2
streamlit==1.47.0
requests==2.31.0
flask==3.0.2
flask-cors==4.0.0
langchain_community==0.3.21